requirements.txt
streamlit>=1.40.0
google-genai>=1.0.0
yt-dlp>=2024.0.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
torch
torchaudio
torchcodec
